removed builtin processors
