+++ tabulator/__init__.py
@@ -6,7 +6,6 @@ from __future__ import unicode_literals
 
 from .stream import Stream
 from . import exceptions
-from . import processors
 
 # Deprecated
 from .topen import topen
