+++ tabulator/helpers.py
@@ -133,24 +133,3 @@ def reset_stream(stream):
         except Exception:
             message = 'Stream is not seekable.'
             raise exceptions.LoadingError(message)
-
-
-def convert_row(row):
-    """Convert row values to python objects.
-    """
-    result = []
-    for value in row:
-        try:
-            if isinstance(value, six.string_types):
-                value = ast.literal_eval(value)
-        except Exception:
-            pass
-        result.append(value)
-    return result
-
-
-def bindify(function):
-    """Add bind method to function.
-    """
-    function.bind = partial(partial, function)
-    return function
